requirements.txt
PyPDF2>=3.0.1
pyvis>=0.3.2
networkx>=3.0
selenium>=4.8.3
python-dotenv>=1.0.0